requirements.txt
git+https://github.com/openai/swarm.git
streamlit 
duckduckgo-search
python-dotenv
